Abstracted read/write logic to separate file
