+++ src/handlers/list.ts
@@ -1,26 +1,15 @@
-import { readFile } from "node:fs/promises";
 import { Task } from "../models/Task";
+import { getTasks } from "../services/tasks";
 
-// it handles list commands
+// It handles list commands that when an user type a command "list",  it prints pending tasks,
+// and "list --all" prints all tasks in Array <Task>
 export namespace ListHandler {
   export type Args = {
     all: boolean;
   };
 
-  const tasksFile = "data/db.json";
-
   export const handle = async (args: Args): Promise<void> => {
-
-    //TODO: Line12-20 make a separate function. You may reuse it
-    let allTasks: Array<Task>;
-    try {
-      const data = await readFile(tasksFile);
-      const stringContent = data.toString();
-      allTasks = JSON.parse(stringContent);
-    } catch (e: unknown) {
-      console.error("Error! We couldn't read the file...");
-      return;
-    }
+    let allTasks: Array<Task> = await getTasks();
 
     const pendingTasks = allTasks.filter((task) => {
       return !task.done;
@@ -29,8 +18,7 @@ export namespace ListHandler {
     // Using the ternary operator
     const tasksToPrint: Array<Task> = args.all ? allTasks : pendingTasks;
 
-    //TODO:line 30-36 make a separate function
-
+    // TODO: Make a separate function
     // map function : changing Array<string> to Array <string> + [] or [x]
     const taskStrings: Array<string> = tasksToPrint.map((task) => {
       if (task.done) {
@@ -40,7 +28,7 @@ export namespace ListHandler {
       }
     });
 
-     //TODO:line 41-43 make a separate function
+    // TODO: Make a separate function
     for (const taskString of taskStrings) {
       console.log(taskString);
     }
